+++ source/EngineGpuKernels/CudaSimulation.cuh
@@ -20,7 +20,12 @@ public:
     void calcCudaTimestep();
 
     void getPixelImage(int2 const& rectUpperLeft, int2 const& rectLowerRight, unsigned char* imageData);
-    void getVectorImage(int2 const& rectUpperLeft, int2 const& rectLowerRight, int2 const& imageSize, double zoom, unsigned char* imageData);
+    void getVectorImage(
+        float2 const& rectUpperLeft,
+        float2 const& rectLowerRight,
+        int2 const& imageSize,
+        double zoom,
+        unsigned char* imageData);
     void getSimulationData(int2 const& rectUpperLeft, int2 const& rectLowerRight, DataAccessTO const& dataTO);
     void setSimulationData(int2 const& rectUpperLeft, int2 const& rectLowerRight, DataAccessTO const& dataTO);
